requirements.txt
fastapi>=0.115.0
uvicorn[standard]>=0.30.0
pydantic>=2.7.0
python-dotenv>=1.0.1
requests>=2.32.0
numpy>=1.26.0
pypdf>=4.2.0
python-multipart>=0.0.9
rank-bm25>=0.2.2
beautifulsoup4>=4.12.0
sentence-transformers>=3.0.0
